backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-multipart==0.0.6
pdfplumber==0.10.0
pytesseract==0.3.10
pillow==10.2.0
pandas==2.2.0
openpyxl==3.1.2
python-dateutil==2.8.2
